+++ ecscluster/variables.tf
@@ -93,9 +93,6 @@ variable "instance_patch_group" {
 variable "ami" {
   type        = string
   description = "AMI to use for cluster instances."
-
-  // If this is left empty, the latest version of the golden ami will be used.
-  default = ""
 }
 
 variable "include_ami_device_names" {
@@ -107,4 +104,14 @@ variable "ami_volumes_delete_on_termination" {
   type        = bool
   description = "Whether to set the `delete_on_termination` flag for volumes included from the AMI."
   default     = true
+}
+
+variable "additional_cloudinit_configs" {
+  type = list(string)
+  description = <<-DESC
+    Additional cloud-init configuration content to be included in the user_data provided to the cluster's autoscaling group. Additional
+    configs are merged with the default cloud-init file using the list(append)+dict(no_replace, recurse_list)+str(append)
+    strategy. See https://cloudinit.readthedocs.io/en/latest/reference/merging.html#built-in-mergers for more informaiton.
+  DESC
+  default = []
 }