+++ newrelic/ec2-agent-cloud-config/variables.tf
@@ -1,0 +1,9 @@
+variable "name_prefix" {
+  type        = string
+  description = "The name prefix to use in the hostname."
+}
+
+variable "new_relic_license_key" {
+  type        = string
+  description = "New Relic License Key."
+}